feat: move fixture_helpers in autonomy/test_tools
