+++ scripts/whitelist.py
@@ -86,3 +86,19 @@ ETHEREUM_ENCRYPTION_PASSWORD  # unused variable (autonomy/test_tools/configurati
 GANACHE_CONFIGURATION  # unused variable (autonomy/test_tools/configurations.py:48)
 HTTP_LOCALHOST  # unused variable (autonomy/test_tools/configurations.py:39)
 KEY_PAIRS  # unused variable (autonomy/test_tools/configurations.py:62)
+UseTendermint  # unused class (autonomy/test_tools/fixture_helpers.py:55)
+_._start_tendermint  # unused method (autonomy/test_tools/fixture_helpers.py:62)
+_.node_address  # unused property (autonomy/test_tools/fixture_helpers.py:81)
+UseFlaskTendermintNode  # unused class (autonomy/test_tools/fixture_helpers.py:87)
+_._start_tendermint  # unused method (autonomy/test_tools/fixture_helpers.py:91)
+_.get_laddr  # unused method (autonomy/test_tools/fixture_helpers.py:120)
+UseGnosisSafeHardHatNet  # unused class (autonomy/test_tools/fixture_helpers.py:129)
+_._start_hardhat  # unused method (autonomy/test_tools/fixture_helpers.py:135)
+gnosis_safe_hardhat_scope_function  # unused variable (autonomy/test_tools/fixture_helpers.py:138)
+hardhat_port  # unused variable (autonomy/test_tools/fixture_helpers.py:138)
+UseGanache  # unused class (autonomy/test_tools/fixture_helpers.py:144)
+_._start_ganache  # unused method (autonomy/test_tools/fixture_helpers.py:150)
+ganache  # unused variable (autonomy/test_tools/fixture_helpers.py:152)
+UseACNNode  # unused class (autonomy/test_tools/fixture_helpers.py:163)
+_._start_acn  # unused method (autonomy/test_tools/fixture_helpers.py:170)
+_._acn_node_image  # unused attribute (autonomy/test_tools/fixture_helpers.py:174)
