+++ tests/test_agents/test_apy_estimation_abci.py
@@ -22,7 +22,8 @@
 import pytest
 from aea.configurations.data_types import PublicId
 
-from tests.fixture_helpers import UseGnosisSafeHardHatNet
+from autonomy.test_tools.fixture_helpers import UseGnosisSafeHardHatNet
+
 from tests.test_agents.base import BaseTestEnd2EndExecution, RoundChecks
