+++ tests/test_agents/test_oracle_abci.py
@@ -24,7 +24,8 @@ from typing import Tuple
 import pytest
 from aea.configurations.data_types import PublicId
 
-from tests.fixture_helpers import UseGnosisSafeHardHatNet
+from autonomy.test_tools.fixture_helpers import UseGnosisSafeHardHatNet
+
 from tests.test_agents.base import BaseTestEnd2EndExecution, RoundChecks
