+++ tests/test_connections/test_abci/test_abci.py
@@ -43,6 +43,7 @@ from hypothesis.strategies import integers
 
 from autonomy.test_tools.configurations import HTTP_LOCALHOST
 from autonomy.test_tools.docker.base import skip_docker_tests
+from autonomy.test_tools.fixture_helpers import UseTendermint
 
 from packages.valory.connections.abci import check_dependencies as dep_utils
 from packages.valory.connections.abci.connection import (
@@ -73,7 +74,6 @@ from packages.valory.protocols.abci.dialogues import AbciDialogue
 from packages.valory.protocols.abci.dialogues import AbciDialogues as BaseAbciDialogues
 
 from tests.conftest import ANY_ADDRESS
-from tests.fixture_helpers import UseTendermint
 from tests.helpers.async_utils import (
     AnotherThreadTask,
     BaseThreadedAsyncLoop,
